MicroBitLog: Introduce online mode, iframe-reload, styles (#111)
Introduces a build step so we can store some reasonable HTML in the
repository and munge it into the C++ array.

The online mode uses the dl.css/js files hosted at microbit.org. They're
live, but we might still change the contract between the page and the
hosted files based on feedback.

The iframe-based reload will reload the page if it detects that it
changes. The process uses postMessage to notify that a reload is needed
based on a simple hash of the contents.