Increase DAPLink I2C wait for data to be ready timeout.
The current values for the retries (5 loops => 5ms) is enough for
normal operation. However, when MakeCode is WebUSB connected to
the micro:bit it must be performing backgroun operations in DAPLink
that might delay how long it takes for operations to be ready.

In the worse case scenarios testing with a lot of serial data +
requesting ADC readings it has sometimes taken up to 8ms for the
combined interrupt signal to be activated by DAPLink with the data
ready. Increasing this value from 5 to 10 shouldn't any other adverse
consequences, other than blocking the fiber in awaitUIPMPacket()
a bit longer.