Ensure header files and .cpp files use the codal namespace appropriately.
Also update the codal-core and codal-nrf52 dependencies as they
include changes on the same topic.

The classes declared in the Mbed compatibility header files have
not been wrapped in the codal namespace, as programmes including
the mbed.h header might not expect them to be inside that namespace.
This header file is not included in any of the codal-microbit-v2
or dependencies, it's mostly used by MakeCode extensions that
were accessing Mbed-specific libraries in micro:bit V1.